Move extensions to their own directory
