+++ WakaTime/Extensions/OptionalExtension.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+extension Optional where Wrapped: Collection {
+    var isEmpty: Bool {
+        self?.isEmpty ?? true
+    }
+}
